feat(service): 重构服务层代码并添加消息队列支持
- 重构了多个服务实现类的代码结构，提高了代码可读性和可维护性
- 在 AdminProductServiceImpl 中添加了消息队列发送逻辑，支持异步处理商品规格相关操作
- 更新了应用配置文件，添加了 RabbitMQ 相关配置
- 将 Message 类重命名为 MessageReturn，更准确地反映了其用途
- 优化了错误处理和消息返回机制，统一使用 MessageReturn 类中的错误信息

+++ parent/service/src/main/java/com/zhuxi/service/Tx/ProductTxService.java
@@ -1,6 +1,7 @@
 package com.zhuxi.service.Tx;
 
-import com.zhuxi.Constant.Message;
+import com.zhuxi.Constant.MessageReturn;
+import com.zhuxi.Exception.MQException;
 import com.zhuxi.Exception.transactionalException;
 import com.zhuxi.mapper.ProductMapper;
 import lombok.extern.slf4j.Slf4j;
@@ -36,16 +37,49 @@ public class ProductTxService {
     }
 
 
+    @Transactional(readOnly = true,propagation = Propagation.SUPPORTS)
+    public List<snowFlakeMap> getSnowFlakeMap(Long lastId,int pageSize){
+        List<snowFlakeMap> snowFlakeMap = productMapper.getSnowFlakeMap(lastId, pageSize);
+        if (snowFlakeMap==null ||snowFlakeMap.isEmpty())
+            throw new MQException(MessageReturn.SELECT_ERROR);
+        return snowFlakeMap;
+    }
+
+
+    @Transactional(readOnly = true,propagation = Propagation.SUPPORTS)
+    public Long getProductSnowFlakeById(Long id) {
+        Long snowFlakeById = productMapper.getProductSnowFlakeById(id);
+        if(snowFlakeById == null)
+            throw new MQException(MessageReturn.SELECT_ERROR);
+        return snowFlakeById;
+    }
+
+    @Transactional(readOnly = true,propagation = Propagation.SUPPORTS)
+    public Long getSpecSnowFlakeById(Long id) {
+        Long snowFlakeById = productMapper.getSpecSnowFlakeById(id);
+        if(snowFlakeById == null)
+            throw new MQException(MessageReturn.SELECT_ERROR);
+        return snowFlakeById;
+    }
+
+    @Transactional(readOnly = true,propagation = Propagation.SUPPORTS)
+    public List<Long> getSpecSnowFlakeByIdList(Long ProductId) {
+        List<Long> snowFlakeById = productMapper.getSpecSnowFlakeByIdList(ProductId);
+        if(snowFlakeById == null || snowFlakeById.isEmpty())
+            throw new MQException(MessageReturn.SELECT_ERROR);
+        return snowFlakeById;
+    }
+
     @Transactional(rollbackFor = transactionalException.class)
     public void putOnSale(Long productId){
 
         boolean existsErrorSpec = productMapper.isExistsErrorSpec(productId);
-        if (existsErrorSpec){
-            throw new transactionalException(Message.EXISTS_ERROR_SPEC);
+        if (!existsErrorSpec){
+            throw new transactionalException(MessageReturn.EXISTS_ERROR_SPEC);
         }
         int i = productMapper.putOnSale(productId);
-        if(i != 0)
-            throw new transactionalException(Message.PUT_SALE_ERROR);
+        if(i != 1)
+            throw new transactionalException(MessageReturn.PUT_SALE_ERROR);
     }
 
 
@@ -53,25 +87,42 @@ public class ProductTxService {
     public Long getProductIdBySnowflakeId(Long snowflake_id){
         Long productId = productMapper.getProductIdBySnowflakeId(snowflake_id);
         if(productId == null)
-            throw new transactionalException(Message.SELECT_ERROR);
+            throw new transactionalException(MessageReturn.SELECT_ERROR);
         return productId;
+    }
+
+    @Transactional(readOnly = true,propagation = Propagation.SUPPORTS)
+    public List<PIdSnowFlake> getSaleProductId(Long lastId,int pageSize){
+        List<PIdSnowFlake> saleProductId = productMapper.getSaleProductId(lastId, pageSize);
+        if(saleProductId == null || saleProductId.isEmpty())
+            throw new transactionalException(MessageReturn.SELECT_ERROR);
+        return saleProductId;
+    }
+
+    @Transactional(readOnly = true,propagation = Propagation.SUPPORTS)
+    public Long getSaleProductIdOne(Long productId){
+        Long saleProductIdOne = productMapper.getSaleProductIdOne(productId);
+        if(saleProductIdOne ==  null)
+            throw new transactionalException(MessageReturn.SELECT_ERROR);
+        return saleProductIdOne;
     }
 
 
     @Transactional(readOnly = true,propagation=Propagation.SUPPORTS)
-    public List<SpecRedisDTO> getSpec(Long lastId,int pageSize){
-        List<SpecRedisDTO> spec = productMapper.getSpec(lastId,pageSize);
-        if(spec.isEmpty())
-            throw new transactionalException(Message.SELECT_ERROR);
+    public List<SpecRedisDTO> getSpec(List<Long> saleProductId){
+        List<SpecRedisDTO> spec = productMapper.getSpec(saleProductId);
+        if(spec==null || spec.isEmpty())
+            throw new transactionalException(MessageReturn.SELECT_ERROR);
 
         return spec;
     }
 
+
     @Transactional(readOnly = true,propagation=Propagation.SUPPORTS)
-    public SpecRedisDTO getSpec(Long productId){
-        SpecRedisDTO spec = productMapper.getSpecOne(productId);
-        if(spec == null)
-            throw new transactionalException(Message.SPEC_NOT_EXIST);
+    public List<SpecRedisDTO> getSpecOne(Long productId){
+        List<SpecRedisDTO> spec = productMapper.getSpecOne(productId);
+        if(spec == null || spec.isEmpty())
+            throw new MQException(MessageReturn.SPEC_NOT_EXIST);
         return spec;
     }
 
@@ -84,7 +135,7 @@ public class ProductTxService {
         }
         ProductDetailVO productDetail = productMapper.getProductDetail(productId);
         if(productDetail == null)
-            throw new transactionalException(Message.SELECT_ERROR);
+            throw new transactionalException(MessageReturn.SELECT_ERROR);
 
         return productDetail;
     }
@@ -101,7 +152,7 @@ public class ProductTxService {
         }
         List<ProductSpecVO> productSpec = productMapper.getProductSpec(productId);
         if(productSpec.isEmpty())
-            throw new transactionalException(Message.SELECT_ERROR);
+            throw new transactionalException(MessageReturn.SELECT_ERROR);
 
         return productSpec;
     }
@@ -111,7 +162,7 @@ public class ProductTxService {
         List<ProductOverviewVO> listProducts = productMapper.getListProductByCreate(dateTime, pageSize);
 
         if (listProducts.isEmpty())
-            throw new transactionalException(Message.SELECT_ERROR);
+            throw new transactionalException(MessageReturn.SELECT_ERROR);
 
         return listProducts;
     }
@@ -121,7 +172,7 @@ public class ProductTxService {
         List<ProductOverviewVO> listProducts = productMapper.getListProductByPriceDESC(lastId,price, pageSize);
 
         if (listProducts.isEmpty())
-            throw new transactionalException(Message.SELECT_ERROR);
+            throw new transactionalException(MessageReturn.SELECT_ERROR);
 
         return listProducts;
     }
@@ -131,7 +182,7 @@ public class ProductTxService {
         List<ProductOverviewVO> listProducts = productMapper.getListProductByPriceASC(lastId,price, pageSize);
 
         if (listProducts.isEmpty())
-            throw new transactionalException(Message.ALREADY_NO_PRODUCT);
+            throw new transactionalException(MessageReturn.ALREADY_NO_PRODUCT);
 
         return listProducts;
     }
@@ -142,10 +193,10 @@ public class ProductTxService {
     }
 
     @Transactional(readOnly = true,propagation = Propagation.SUPPORTS)
-    public ProductDetailVO getListProduct(Long lastId){
-        ProductDetailVO product = productMapper.getListProduct(lastId);
+    public ProductDetailVO getListProductMQ(Long lastId){
+        ProductDetailVO product = productMapper.getListProductMQ(lastId);
         if(product == null)
-            throw new transactionalException(Message.PRODUCT_ID_NOT_EXIST);
+            throw new MQException(MessageReturn.PRODUCT_ID_NOT_EXIST);
 
         return  product;
     }
@@ -154,7 +205,7 @@ public class ProductTxService {
     public List<AdminProductVO> getListAdminProductsDESC(Long lastId, Integer pageSize){
         List<AdminProductVO> listAdminProductsDESC = productMapper.getListAdminProductsDESC(lastId, pageSize);
         if(listAdminProductsDESC.isEmpty())
-            throw new transactionalException(Message.SELECT_ERROR);
+            throw new transactionalException(MessageReturn.SELECT_ERROR);
 
         return listAdminProductsDESC;
     }
@@ -163,7 +214,7 @@ public class ProductTxService {
     public List<AdminProductVO> getListAdminProductsASC(Long lastId,Integer pageSize){
         List<AdminProductVO> listAdminProductsASC = productMapper.getListAdminProductsASC(lastId, pageSize);
         if(listAdminProductsASC.isEmpty())
-            throw new transactionalException(Message.SELECT_ERROR);
+            throw new transactionalException(MessageReturn.SELECT_ERROR);
 
         return listAdminProductsASC;
     }
@@ -172,7 +223,7 @@ public class ProductTxService {
     public List<ProductSpecDetailVO> getProductSpecDetail(Long productId){
         List<ProductSpecDetailVO> productSpecDetail = productMapper.getProductSpecDetail(productId);
         if(productSpecDetail.isEmpty())
-            throw new transactionalException(Message.SELECT_ERROR);
+            throw new transactionalException(MessageReturn.SELECT_ERROR);
 
         return productSpecDetail;
     }
@@ -181,7 +232,7 @@ public class ProductTxService {
     public Integer getRealStock(Long productId, Long specId){
         Integer realStock = productMapper.getRealStock(productId, specId);
         if(realStock < 0)
-            throw new transactionalException(Message.SELECT_ERROR);
+            throw new transactionalException(MessageReturn.SELECT_ERROR);
 
         return realStock;
     }
@@ -189,24 +240,24 @@ public class ProductTxService {
     @Transactional(rollbackFor = transactionalException.class)
     public void addBasePics(String coverUrl,List<String> images,Long productId){
         if(productMapper.addBasePics(coverUrl,images,productId) < 0)
-            throw new transactionalException(Message.INSERT_ERROR);
+            throw new transactionalException(MessageReturn.INSERT_ERROR);
     }
 
     @Transactional(rollbackFor = transactionalException.class)
     public void addSpecCoverUrl(String coverUrl,Long productID,Long id){
         if(productMapper.addSpecCoverUrl(coverUrl,productID,id) < 0)
-            throw new transactionalException(Message.INSERT_ERROR);
+            throw new transactionalException(MessageReturn.INSERT_ERROR);
     }
 
     @Transactional(rollbackFor = transactionalException.class)
     public Integer updateBase(ProductBaseUpdateDTO productBaseUpdateDTO){
 
         if(productMapper.updateProductBase(productBaseUpdateDTO) <= 0)
-            throw new transactionalException(Message.UPDATE_ERROR);
+            throw new transactionalException(MessageReturn.UPDATE_ERROR);
 
         Integer productStatus = productMapper.getProductStatus(productBaseUpdateDTO.getId());
         if(productStatus == null)
-            throw new transactionalException(Message.NO_PRODUCT);
+            throw new transactionalException(MessageReturn.NO_PRODUCT);
         return productStatus;
     }
 
@@ -222,7 +273,7 @@ public class ProductTxService {
             if(batchResult.getUpdateCounts() != null){
                 for (int updateCount : batchResult.getUpdateCounts())
                     if(updateCount <= 0)
-                        throw new transactionalException(Message.UPDATE_ERROR);
+                        throw new transactionalException(MessageReturn.UPDATE_ERROR);
             }
         }
     }
@@ -231,7 +282,7 @@ public class ProductTxService {
     @Transactional(rollbackFor = transactionalException.class)
     public void addBase(ProductBaseDTO productBaseDTO){
         if(!productMapper.addBase(productBaseDTO))
-            throw new transactionalException(Message.INSERT_ERROR);
+            throw new transactionalException(MessageReturn.INSERT_ERROR);
     }
 
     // 添加商品规格信息
@@ -247,7 +298,7 @@ public class ProductTxService {
                 if(batchResult.getUpdateCounts() != null){
                     for (int updateCount : batchResult.getUpdateCounts())
                         if(updateCount <= 0)
-                            throw new transactionalException(Message.INSERT_ERROR);
+                            throw new transactionalException(MessageReturn.INSERT_ERROR);
                 }
             }
     }
@@ -255,15 +306,18 @@ public class ProductTxService {
     @Transactional(rollbackFor = transactionalException.class)
     public void addRealStock(List<RealStockDTO> realStockDTO){
         if(!productMapper.addRealStock(realStockDTO))
-            throw new transactionalException(Message.INSERT_ERROR);
+            throw new transactionalException(MessageReturn.INSERT_ERROR);
     }
 
     @Transactional(rollbackFor = transactionalException.class)
     public void delete(Long id){
+        if (productMapper.deleteProductRealStock(id) <= 0){
+            throw new transactionalException(MessageReturn.DELETE_ERROR);
+        }
         if (productMapper.deleteProductSpec(id) <= 0){
-            throw new transactionalException(Message.DELETE_ERROR);
+            throw new transactionalException(MessageReturn.DELETE_ERROR);
         }
         if(productMapper.deleteProductBase(id) <= 0)
-            throw new transactionalException(Message.DELETE_ERROR);
+            throw new transactionalException(MessageReturn.DELETE_ERROR);
     }
 }
